Updated NuGet packages
The solution was refusing to build until I updated Fody and Costura ("Could not load file or assembly 'Microsoft.Build.Framework, ...' or one of its dependencies. The system cannot find the file specified..."), but given that I have no idea what weaving is, there's a chance I accidentally broke something in the process.